// src/app/resolvingoferrors/models/unsolicitedactionreport.ts
export interface unsolicitedactionreport {

   
    Link:string;
    ResolveType: string;
    TelephoneNumber: string;
    TransactionID: string;
    ResolveRemarks: string;
    CreatedBy: string;
    CreatedOn: string;
    Duration: string;
    Source: string;
    Status: string;
    TransactionCommand: string;
}